move/rename files + update example group name with convention
