+++ modules/appeals_api/spec/requests/v2/decision_reviews/supplemental_claims/evidence_submissions_spec.rb
@@ -3,13 +3,12 @@
 require 'rails_helper'
 require AppealsApi::Engine.root.join('spec', 'spec_helper.rb')
 
-describe AppealsApi::V1::DecisionReviews::NoticeOfDisagreements::EvidenceSubmissionsController, type: :request do
+Rspec.describe 'AppealsApi::V2::DecisionReviews::SupplementalClaims::EvidenceSubmissions', type: :request do
   include FixtureHelpers
-  let(:notice_of_disagreement) { create(:notice_of_disagreement, :board_review_hearing) }
-  let(:headers) { fixture_as_json 'decision_reviews/v1/valid_10182_headers.json' }
-  let(:evidence_submissions) { create_list(:evidence_submission, 3, supportable: notice_of_disagreement) }
-  let(:path) { '/services/appeals/v1/decision_reviews/notice_of_disagreements/evidence_submissions/' }
-
+  let(:supplemental_claim) { create(:supplemental_claim) }
+  let(:headers) { fixture_as_json 'decision_reviews/v2/valid_200995_headers.json' }
+  let(:evidence_submissions) { create_list(:evidence_submission, 3, supportable: supplemental_claim) }
+  let(:path) { '/services/appeals/v2/decision_reviews/supplemental_claims/evidence_submissions/' }
   let(:parsed) { JSON.parse(response.body) }
 
   def stub_upload_location(expected_location = 'http://some.fakesite.com/path/uuid')
@@ -17,59 +16,49 @@ describe AppealsApi::V1::DecisionReviews::NoticeOfDisagreements::EvidenceSubmiss
   end
 
   describe '#create' do
-    context 'when corresponding notice of disagreement record not found' do
+    context 'when corresponding supplemental claim record not found' do
       it 'returns an error' do
         stub_upload_location
-        post(path, params: { nod_uuid: 1979 }, headers:)
+        post(path, params: { sc_uuid: 1979 }, headers:)
 
-        expect(response.status).to eq 404
-        expect(response.body).to include 'not found'
+        expect(response).to have_http_status :not_found
+        expect(response.body).to include 'SupplementalClaim with uuid 1979 not found'
       end
     end
 
-    context 'when corresponding notice of disagreement record found' do
-      it "returns an error if nod 'boardReviewOption' is not 'evidence_submission'" do
-        stub_upload_location
-        post(path, params: { nod_uuid: notice_of_disagreement.id }, headers:)
-
-        expect(response.status).to eq 422
-        expect(response.body).to include "'boardReviewOption' must be 'evidence_submission'"
-      end
-
-      context "when nod record 'auth_headers' are present" do
+    context 'when corresponding supplemental claim record found' do
+      context "when sc record 'auth_headers' are present" do
         it 'returns success with 202' do
           stub_upload_location
-          notice_of_disagreement.update(board_review_option: 'evidence_submission')
-          post(path, params: { nod_uuid: notice_of_disagreement.id }, headers:)
+          post(path, params: { sc_uuid: supplemental_claim.id }, headers:)
 
-          expect(response.status).to eq 202
-          expect(response.body).to include notice_of_disagreement.id
+          expect(response).to have_http_status :accepted
+          expect(response.body).to include supplemental_claim.id
         end
 
-        it "returns an error if request 'headers['X-VA-SSN'] and NOD record SSNs do not match" do
+        it "returns an error if request 'headers['X-VA-SSN'] and SC record SSNs do not match" do
           stub_upload_location
-          notice_of_disagreement.update(board_review_option: 'evidence_submission')
           headers['X-VA-SSN'] = '1111111111'
-          post(path, params: { nod_uuid: notice_of_disagreement.id }, headers:)
+          post(path, params: { sc_uuid: supplemental_claim.id }, headers:)
 
-          expect(response.status).to eq 422
+          expect(response).to have_http_status :unprocessable_entity
           expect(response.body).to include "'X-VA-SSN' does not match"
         end
       end
 
-      context "when nod record 'auth_headers' are not present" do
+      context "when sc record 'auth_headers' are not present" do
         # if PII expunged not validating for matching SSNs
         it 'creates the evidence submission and returns upload location' do
           stub_upload_location 'http://another.fakesite.com/rewrittenpath/uuid'
-          notice_of_disagreement.update(board_review_option: 'evidence_submission', auth_headers: nil)
-          post(path, params: { nod_uuid: notice_of_disagreement.id }, headers:)
+          supplemental_claim.auth_headers = nil
+          post(path, params: { sc_uuid: supplemental_claim.id }, headers:)
 
           data = JSON.parse(response.body)['data']
           expect(data).to have_key('id')
           expect(data).to have_key('type')
           expect(data['attributes']['status']).to eq('pending')
-          expect(data['attributes']['appealId']).to eq(notice_of_disagreement.id)
-          expect(data['attributes']['appealType']).to eq('NoticeOfDisagreement')
+          expect(data['attributes']['appealId']).to eq(supplemental_claim.id)
+          expect(data['attributes']['appealType']).to eq('SupplementalClaim')
           expect(data['attributes']['location']).to eq('http://another.fakesite.com/rewrittenpath/uuid')
         end
       end
@@ -78,26 +67,25 @@ describe AppealsApi::V1::DecisionReviews::NoticeOfDisagreements::EvidenceSubmiss
         allow_any_instance_of(VBADocuments::UploadSubmission).to(
           receive(:get_location).and_raise('Unable to provide document upload location')
         )
-        notice_of_disagreement.update(board_review_option: 'evidence_submission', auth_headers: nil)
-        post(path, params: { nod_uuid: notice_of_disagreement.id }, headers:)
+        supplemental_claim.auth_headers = nil
+        post(path, params: { sc_uuid: supplemental_claim.id }, headers:)
 
-        expect(response.status).to eq 500
+        expect(response).to have_http_status :internal_server_error
         expect(response.body).to include('Unable to provide document upload location')
       end
     end
 
-    it "returns an error when 'nod_uuid' parameter is missing" do
+    it "returns an error when 'sc_uuid' parameter is missing" do
       stub_upload_location
       post(path, headers:)
 
-      expect(response.status).to eq 400
-      expect(response.body).to include 'Must supply a corresponding NOD'
+      expect(response).to have_http_status :bad_request
+      expect(response.body).to include 'Must supply a corresponding Supplemental Claim'
     end
 
     it 'stores the source from headers' do
       stub_upload_location
-      notice_of_disagreement.update(board_review_option: 'evidence_submission')
-      post(path, params: { nod_uuid: notice_of_disagreement.id }, headers:)
+      post(path, params: { sc_uuid: supplemental_claim.id }, headers:)
       data = JSON.parse(response.body)['data']
       record = AppealsApi::EvidenceSubmission.find_by(guid: data['id'])
       expect(record.source).to eq headers['X-Consumer-Username']
@@ -124,20 +112,20 @@ describe AppealsApi::V1::DecisionReviews::NoticeOfDisagreements::EvidenceSubmiss
 
     it 'returns details for the evidence submission' do
       es = evidence_submissions.sample
-      nod_uuid = es.supportable_id
+      sc_uuid = es.supportable_id
       get "#{path}#{es.guid}"
       submission = JSON.parse(response.body)['data']
 
       expect(submission['id']).to eq es.guid
       expect(submission['type']).to eq('evidenceSubmission')
       expect(submission['attributes']['status']).to eq('pending')
-      expect(submission['attributes']['appealId']).to eq(nod_uuid)
-      expect(submission['attributes']['appealType']).to eq('NoticeOfDisagreement')
+      expect(submission['attributes']['appealId']).to eq(sc_uuid)
+      expect(submission['attributes']['appealType']).to eq('SupplementalClaim')
     end
 
     it 'returns an error if record is not found' do
       get "#{path}/bueller"
-      expect(response.status).to eq 404
+      expect(response).to have_http_status :not_found
       expect(response.body).to include 'Record not found'
     end
   end
